requirements.txt
torch
Pillow
matplotlib
scipy
python-dotenv
torchvision